HW4 tasks 1,2,3,4,5,6 completed - Added freepmem() syscall to report free physical memory. - Modified sys_sbrk() for lazy allocation. - Added page fault handler in usertrap() to allocate pages. - Updated uvmunmap() to skip unmapped pages and prevent panic. - Extended malloc/morecore to use 64-bit size types for full address space. - Verified behavior with memory-user test (cases 1–3) and free command.

+++ user/umalloc.c
@@ -11,7 +11,7 @@ typedef long Align;
 union header {
   struct {
     union header *ptr;
-    uint size;
+    uint64 size;
   } s;
   Align x;
 };
@@ -44,7 +44,7 @@ free(void *ap)
 }
 
 static Header*
-morecore(uint nu)
+morecore(uint64 nu)
 {
   char *p;
   Header *hp;
@@ -61,10 +61,10 @@ morecore(uint nu)
 }
 
 void*
-malloc(uint nbytes)
+malloc(uint64 nbytes)
 {
   Header *p, *prevp;
-  uint nunits;
+  uint64 nunits;
 
   nunits = (nbytes + sizeof(Header) - 1)/sizeof(Header) + 1;
   if((prevp = freep) == 0){
